renamed helpers to utils in React side
